BackEnd/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-CORS==3.0.10
Flask-RESTful==0.3.9
python-dotenv==1.0.0 